Move InfoModal to its own module
