+++ client/account/tokens/AccountTokensPage.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import { Button, ButtonGroup, Modal } from '@rocket.chat/fuselage';
 
 import { useTranslation } from '../../contexts/TranslationContext';
 import { useEndpointDataExperimental } from '../../hooks/useEndpointDataExperimental';
@@ -9,23 +8,6 @@ import Page from '../../components/basic/Page';
 import AccountTokensTable from './AccountTokensTable';
 import AddToken from './AddToken';
 import NotAuthorizedPage from '../../components/NotAuthorizedPage';
-
-export const InfoModal = ({ title, content, icon, onConfirm, onClose, confirmText, cancelText, ...props }) => <Modal {...props}>
-	<Modal.Header>
-		{icon}
-		<Modal.Title>{title}</Modal.Title>
-		<Modal.Close onClick={onClose}/>
-	</Modal.Header>
-	<Modal.Content fontScale='p1'>
-		{content}
-	</Modal.Content>
-	<Modal.Footer>
-		<ButtonGroup align='end'>
-			{cancelText && <Button onClick={onClose}>{cancelText}</Button>}
-			{confirmText && onConfirm && <Button primary onClick={onConfirm}>{confirmText}</Button>}
-		</ButtonGroup>
-	</Modal.Footer>
-</Modal>;
 
 const emptyObj = {};
